refactor(866): move GoogleIntegration jobs to their sections (G5 lane B1)
Four jobs leave `Humans.Infrastructure/Jobs`, following lane 5b-1's pattern: the job
lands in its owning section's `Contracts/` folder, `public`, taking the folder's
namespace. No registration site moved, no job id changed, no `retired` entry needed.

Ownership was re-measured from each body rather than the filename:

- `DriveActivityMonitorJob` -> `Humans.Monitor/Contracts/`. The only service it drives is
  `IDriveActivityMonitorService`, which Monitor owns; it names nothing from
  GoogleIntegration. `Humans.Monitor.csproj` already said so in as many words. This
  matches the endstate inventory's row, which assigned it to Monitor all along.
- `GoogleResourceReconciliationJob`, `ProcessGoogleSyncOutboxJob`,
  `GoogleResourceProvisionJob` -> `Humans.GoogleIntegration/Contracts/`.

`public` under `Contracts/` is forced, not chosen: Shell names each concrete type at two
sites (`AddScoped<T>()` and the recurring roll-call) and HUM0034 makes any other public
type in a section assembly a build error.

`GoogleResourceProvisionJob` still has no enqueue site anywhere in `src/` - the
2026-08-05 debt-ledger entry on it is open and undecided. The move keeps it exactly as
dead as it was; the ledger entry's path is corrected in the same commit.

One package reference added: `Hangfire.Core` on `Humans.Monitor` for
`[DisableConcurrentExecution]`. Package, not project, so no cycle is possible, and it was
already arriving transitively. Zero new `<ProjectReference>`; one removed -
`Humans.Infrastructure` -> `Humans.Monitor.Contracts` existed only for the job that left.

Comment and doc drift the move caused is corrected alongside: Monitor's section doc and
freshness triggers, the Monitor/Teams/Notifications contracts-leaf csproj rationales,
`IGoogleSyncOutboxProcessor`'s public-surface remark, and the Google Integration feature
doc's triggers.

Refs nobodies-collective/Humans#866
