+++ docs/architecture/debt-ledger.yml
@@ -291,7 +291,7 @@ inbox:
     what: "GoogleController.SyncExecute / SyncExecuteAll run a full Workspace reconcile inline on the request thread. Moving them to Hangfire (the TicketController.Sync pattern) is blocked on UI: Views/Google/Sync.cshtml:142,163 calls loadTab() in the fetch continuation, so the tab refresh depends on the POST having completed the reconcile — fire-and-forget would re-render stale diffs. Needs a progress/polling affordance (deferred from nobodies-collective/Humans#950)"
     review: panel
   - added: 2026-08-05
-    what: "GoogleResourceProvisionJob (src/Humans.Infrastructure/Jobs/GoogleResourceProvisionJob.cs) has no enqueue site anywhere in src — no BackgroundJob.Enqueue, no RecurringJob.AddOrUpdate, no reference outside its own declaration. It calls IGoogleSyncService.ProvisionTeamFolderAsync (a Drive write), so it is not inert if something ever schedules it. Decide whether team-folder provisioning is meant to run through it or through the Google sync outbox, then wire it or delete it (found while auditing nobodies-collective/Humans#950)"
+    what: "GoogleResourceProvisionJob (src/Sections/Humans.GoogleIntegration/Contracts/GoogleResourceProvisionJob.cs, moved there from src/Humans.Infrastructure/Jobs at the G5 jobs move and still undecided) has no enqueue site anywhere in src — no BackgroundJob.Enqueue, no RecurringJob.AddOrUpdate, no reference outside its own declaration. It calls IGoogleSyncService.ProvisionTeamFolderAsync (a Drive write), so it is not inert if something ever schedules it. Decide whether team-folder provisioning is meant to run through it or through the Google sync outbox, then wire it or delete it (found while auditing nobodies-collective/Humans#950)"
     review: panel
   - added: 2026-08-05
     what: "GitHubGuideContentSource.GetMarkdownAsync / ListMarkdownStemsAsync accept a CancellationToken they never forward — Octokit's GetRawContentByRef / GetAllContentsByRef expose no CT overload. Read-only so harmless, but the signature promises something it cannot deliver (found while auditing nobodies-collective/Humans#950)"
