+++ src/Humans.Web/Extensions/Infrastructure/GoogleWorkspaceInfrastructureExtensions.cs
@@ -1,6 +1,7 @@
 using Humans.Application.Configuration;
+using Humans.GoogleIntegration.Contracts;
 using Humans.Infrastructure.Configuration;
-using Humans.Infrastructure.Jobs;
+using Humans.Monitor.Contracts;
 
 namespace Humans.Web.Extensions.Infrastructure;
 
@@ -20,9 +21,11 @@ namespace Humans.Web.Extensions.Infrastructure;
 /// section reads the same two configuration keys to decide which of its own connector sets to
 /// bind.
 /// <para>
-/// The three recurring jobs stay in <c>Humans.Infrastructure/Jobs</c> because
-/// <c>UseHumansRecurringJobs</c> names them by concrete type (step 6b); each reaches the
-/// section through <c>Humans.GoogleIntegration.Contracts</c>.
+/// The three recurring jobs now live in their owning sections' <c>Contracts/</c> folders —
+/// two in <c>Humans.GoogleIntegration</c>, and <c>DriveActivityMonitorJob</c> in
+/// <c>Humans.Monitor</c>, whose service it is the only caller of. The registrations stay here
+/// because <c>UseHumansRecurringJobs</c> and this method both name them by concrete type,
+/// which is also why they are public rather than internal (HUM0034).
 /// </para>
 /// </remarks>
 internal static class GoogleWorkspaceInfrastructureExtensions
