+++ src/Sections/Humans.GoogleIntegration/Contracts/ProcessGoogleSyncOutboxJob.cs
@@ -1,9 +1,8 @@
 using Hangfire;
 using Microsoft.Extensions.Logging;
 using Humans.Application.Interfaces;
-using Humans.GoogleIntegration.Contracts;
 
-namespace Humans.Infrastructure.Jobs;
+namespace Humans.GoogleIntegration.Contracts;
 
 /// <summary>
 /// Runs the Google sync outbox drain every 10 minutes via Hangfire. The queue semantics —
@@ -11,6 +10,12 @@ namespace Humans.Infrastructure.Jobs;
 /// user GoogleEmailStatus mirror — live inside the GoogleIntegration section behind
 /// <see cref="IGoogleSyncOutboxProcessor"/>; this job is the scheduler shim around it.
 /// </summary>
+/// <remarks>
+/// Moved out of <c>Humans.Infrastructure/Jobs</c> at the G5 jobs move
+/// (nobodies-collective/Humans#866). Public and under <c>Contracts/</c> because Shell names the
+/// concrete type at two sites (<c>AddScoped</c> and the recurring roll-call) and HUM0034 allows
+/// a section's public types only there.
+/// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public class ProcessGoogleSyncOutboxJob(
     IGoogleSyncOutboxProcessor outbox,
